+++ src/protocolgame.cpp
@@ -2479,7 +2479,7 @@ void ProtocolGame::AddPlayerStats(NetworkMessage& msg)
 		msg.addByte(
 		    static_cast<uint8_t>(std::min<uint32_t>(player->getBaseMagicLevel(), std::numeric_limits<uint8_t>::max())));
 	}
-	msg.addByte(static_cast<uint8_t>(player->getMagicLevelPercent()));
+	msg.addByte(player->getMagicLevelPercent());
 
 	msg.addByte(player->getSoul());
 
@@ -2509,13 +2509,13 @@ void ProtocolGame::AddPlayerSkills(NetworkMessage& msg)
 		for (uint8_t i = SKILL_FIRST; i <= SKILL_LAST; ++i) {
 			msg.addByte(
 			    std::min<uint8_t>(static_cast<uint8_t>(player->getSkillLevel(i)), std::numeric_limits<uint8_t>::max()));
-			msg.addByte(static_cast<uint8_t>(player->getSkillPercent(i)));
+			msg.addByte(player->getSkillPercent(i));
 		}
 	} else {
 		for (uint8_t i = SKILL_FIRST; i <= SKILL_LAST; ++i) {
 			msg.add<uint16_t>(std::min<uint16_t>(player->getSkillLevel(i), std::numeric_limits<uint16_t>::max()));
 			msg.add<uint16_t>(player->getBaseSkill(i));
-			msg.addByte(static_cast<uint8_t>(player->getSkillPercent(i)));
+			msg.addByte(player->getSkillPercent(i));
 		}
 
 		for (uint8_t i = SPECIALSKILL_FIRST; i <= SPECIALSKILL_LAST; ++i) {
